add builtin execution

+++ srcs/ft_env.c
@@ -6,13 +6,13 @@
 /*   By: naharagu <naharagu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/28 22:40:44 by naharagu          #+#    #+#             */
-/*   Updated: 2023/03/20 23:04:44 by naharagu         ###   ########.fr       */
+/*   Updated: 2023/03/24 21:42:35 by naharagu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
 
-int	ft_env(int argc, char **argv, char **env)
+int	ft_env(size_t argc, char **argv, char **env)
 {
 	int	i;
