+++ srcs/ft_pwd.c
@@ -6,13 +6,13 @@
 /*   By: naharagu <naharagu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/28 22:40:44 by naharagu          #+#    #+#             */
-/*   Updated: 2023/03/20 23:06:05 by naharagu         ###   ########.fr       */
+/*   Updated: 2023/03/24 21:41:31 by naharagu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
 
-int	ft_pwd(int argc, char **argv)
+int	ft_pwd(size_t argc, char **argv)
 {
 	char	*pwd;
